refac: add style formatting for all files in src folder (#71)

+++ src/pages/OnCall.vue
@@ -1,24 +1,17 @@
 <template>
   <h1>
     {{ t('OnCall') }}
-    <information-dialog 
-      :title="t('OnCall')"
-      :info="[{title:'', info: t('OnCallInfo')}]"
-    />
-    
+    <information-dialog :title="t('OnCall')" :info="[{title: '', info: t('OnCallInfo')}]" />
+
     <v-btn
       perms="write:oncalls"
       prepend-icon="add"
       class="no-cap-btn bg-primary-600"
-      style="position: absolute; right: 10px;"
+      style="position: absolute; right: 10px"
       :text="t('AddOnCall')"
       @click="dialog = true"
     />
-    <on-call-add 
-      :dialog="dialog"
-      :item="selectedItem"
-      @close="close"
-    />
+    <on-call-add :dialog="dialog" :item="selectedItem" @close="close" />
   </h1>
 
   <v-data-table-server
@@ -35,15 +28,12 @@
     sort-asc-icon="arrow_drop_up"
     @update:options="setPagination"
   >
-    <template 
-      v-for="desc in ['repeatDays', 'repeatWeeks', 'repeatMonths']"
-      #[`item.${desc}`]="{ item }"
-    >
+    <template v-for="desc in ['repeatDays', 'repeatWeeks', 'repeatMonths']" #[`item.${desc}`]="{item}">
       {{ item[desc as 'repeatDays']?.join(', ') ?? '' }}
     </template>
-    <template #[`item.usersEmails`]="{ item }">
+    <template #[`item.usersEmails`]="{item}">
       <v-chip
-        v-for="email in item.usersEmails.filter((e) => emails[e] !== undefined)"
+        v-for="email in item.usersEmails.filter(e => emails[e] !== undefined)"
         :key="email"
         class="chip"
         size="x-small"
@@ -51,7 +41,7 @@
         :text="emails[email]"
       />
       <v-chip
-        v-for="email in item.usersEmails.filter((e) => emails[e] === undefined)"
+        v-for="email in item.usersEmails.filter(e => emails[e] === undefined)"
         :key="email"
         class="chip critical"
         size="x-small"
@@ -59,14 +49,13 @@
         :text="email"
       />
     </template>
-    <template #[`item.groupIds`]="{ item }">
+    <template #[`item.groupIds`]="{item}">
       {{ getGroupNames(item.groupIds).join(', ') }}
     </template>
-    <template 
-      v-for="desc in ['startTime', 'endTime']"
-      #[`item.${desc}`]="{item}"
-    >
-      {{ item[desc as 'startTime' | 'endTime'] ? filters.hhmmUtcToLocal(item[desc as 'startTime' | 'endTime']!) : null }}
+    <template v-for="desc in ['startTime', 'endTime']" #[`item.${desc}`]="{item}">
+      {{
+        item[desc as 'startTime' | 'endTime'] ? filters.hhmmUtcToLocal(item[desc as 'startTime' | 'endTime']!) : null
+      }}
     </template>
     <template #[`item.actions`]="{item}">
       <v-btn
@@ -95,59 +84,57 @@
 </template>
 
 <script lang="ts" setup>
-import { useFilters } from '@/filters';
-import type { Store } from '@/plugins/store/types';
-import type { Pagination } from '@/plugins/store/types/alerts-types';
-import type { OnCall } from '@/plugins/store/types/onCall-types';
-import { computed, ref, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useStore } from 'vuex';
+import {useFilters} from '@/filters'
+import type {Store} from '@/plugins/store/types'
+import type {Pagination} from '@/plugins/store/types/alerts-types'
+import type {OnCall} from '@/plugins/store/types/onCall-types'
+import {computed, ref, watch} from 'vue'
+import {useI18n} from 'vue-i18n'
+import {useStore} from 'vuex'
 
 definePage({
   meta: {
-    title: "On Call",
+    title: 'On Call',
     requiresAuth: true
   }
-});
+})
 
-
-const { t } = useI18n()
+const {t} = useI18n()
 const store: Store = useStore()
 const filters = useFilters()
 
 const dialog = ref(false)
 const selectedItem = ref<OnCall | null>(null)
 
-const headers = ref<{title: string, key: keyof OnCall | 'actions', info?: string | string[], align?: 'end'}[]>([
-  { title: t('Users'), key: 'usersEmails'},
-  { title: t('Groups'), key: 'groupIds'},
-  { title: t('Start'), key: 'startTime', },
-  { title: t('End'), key: 'endTime' },
-  { title: t('StartDate'), key: 'startDate' },
-  { title: t('EndDate'), key: 'endDate' },
-  { title: t('Days'), key: 'repeatDays' },
-  { title: t('Weeks'), key: 'repeatWeeks' },
-  { title: t('Months'), key: 'repeatMonths' },
-  { title: t('Actions'), key: 'actions', align: 'end' },
+const headers = ref<{title: string; key: keyof OnCall | 'actions'; info?: string | string[]; align?: 'end'}[]>([
+  {title: t('Users'), key: 'usersEmails'},
+  {title: t('Groups'), key: 'groupIds'},
+  {title: t('Start'), key: 'startTime'},
+  {title: t('End'), key: 'endTime'},
+  {title: t('StartDate'), key: 'startDate'},
+  {title: t('EndDate'), key: 'endDate'},
+  {title: t('Days'), key: 'repeatDays'},
+  {title: t('Weeks'), key: 'repeatWeeks'},
+  {title: t('Months'), key: 'repeatMonths'},
+  {title: t('Actions'), key: 'actions', align: 'end'}
 ])
 const items = computed(() => store.state.onCalls.items)
-const computedHeaders = computed(() => headers.value.filter(h => store.state.config.customer_views ? true : h.key != 'customer'))
-const emails = computed(() => 
-  Object.fromEntries(store.state.users.emails.map(e => [e.email, e.name]))
+const computedHeaders = computed(() =>
+  headers.value.filter(h => (store.state.config.customer_views ? true : h.key != 'customer'))
 )
+const emails = computed(() => Object.fromEntries(store.state.users.emails.map(e => [e.email, e.name])))
 
 const pagination = computed({
-  get:() => store.getters['onCalls/pagination'],
-  set: (value) => store.dispatch('onCalls/setPagination', value)
+  get: () => store.getters['onCalls/pagination'],
+  set: value => store.dispatch('onCalls/setPagination', value)
 })
 
 const refresh = computed(() => store.state.refresh)
 
-watch(refresh, (val) => {
+watch(refresh, val => {
   if (!val) return
   getItems()
 })
-
 
 function getItems() {
   store.dispatch('onCalls/getOnCalls')
@@ -167,36 +154,33 @@ function getGroups() {
 }
 
 function deleteItem(item: OnCall) {
-  confirm(t('ConfirmDelete')) &&
-  store.dispatch('onCalls/deleteOnCall', item.id!)
+  confirm(t('ConfirmDelete')) && store.dispatch('onCalls/deleteOnCall', item.id!)
   getItems()
 }
 
 function editItem(item: OnCall) {
-  selectedItem.value =  item
+  selectedItem.value = item
   dialog.value = true
 }
 
-
 function copyItem(item: OnCall) {
-  selectedItem.value =  {...item, id: undefined}
+  selectedItem.value = {...item, id: undefined}
   dialog.value = true
 }
 
 function getGroupNames(ids: string[]) {
-  return ids.map((id) => getGroupName(id))
+  return ids.map(id => getGroupName(id))
 }
 
 function getGroupName(id: string) {
-    const group = store.state.notificationGroups.items.find((u) => u.id == id)
-    return group?.name ?? id
+  const group = store.state.notificationGroups.items.find(u => u.id == id)
+  return group?.name ?? id
 }
 
 function close() {
   dialog.value = false
   selectedItem.value = null
 }
-
 
 getItems()
 getEmails()
