+++ src/pages/Perms.vue
@@ -5,32 +5,25 @@
     :label="t('Search')"
     clearable
     hide-details
-    style="position: absolute; top: 2.5px;right: calc(25vw); width: 40vw; background: white;"
+    style="position: absolute; top: 2.5px; right: calc(25vw); width: 40vw; background: white"
   />
   <v-row>
     <v-col cols="auto">
       <h1>
         {{ t('Permissions') }}
-        
+
         <v-btn
           v-has-perms="'write:perms'"
           prepend-icon="add"
           class="no-cap-btn bg-primary-600"
-          style="position: absolute; right: 10px;"
+          style="position: absolute; right: 10px"
           :text="t('AddPermission')"
           @click="dialog = true"
         />
-        <perm-add 
-          :dialog="dialog"
-          :item="selectedItem"
-          @close="close"
-        />
+        <perm-add :dialog="dialog" :item="selectedItem" @close="close" />
       </h1>
     </v-col>
-    <v-col
-      cols="auto"
-      align-self="center"
-    >
+    <v-col cols="auto" align-self="center">
       <perms-filter />
     </v-col>
   </v-row>
@@ -50,20 +43,10 @@
   >
     <template #[`item.match`]="{item}">
       {{ item.match }}
-      <v-icon
-        v-if="isDefault(item.match)" 
-        icon="lock"
-      />
+      <v-icon v-if="isDefault(item.match)" icon="lock" />
     </template>
     <template #[`item.scopes`]="{item}">
-      <v-chip
-        v-for="scope in item.scopes"
-        :key="scope"
-        outline
-        variant="flat"
-        size="small"
-        class="chip"
-      >
+      <v-chip v-for="scope in item.scopes" :key="scope" outline variant="flat" size="small" class="chip">
         {{ scope }}
       </v-chip>
     </template>
@@ -96,22 +79,21 @@
 </template>
 
 <script lang="ts" setup>
-import type { Store } from '@/plugins/store/types';
-import type { SortBy } from '@/plugins/store/types/alerts-types';
-import type { Permission } from '@/plugins/store/types/perms-types';
-import { computed, ref, watch } from 'vue';
-import { useI18n } from 'vue-i18n';
-import { useStore } from 'vuex';
+import type {Store} from '@/plugins/store/types'
+import type {SortBy} from '@/plugins/store/types/alerts-types'
+import type {Permission} from '@/plugins/store/types/perms-types'
+import {computed, ref, watch} from 'vue'
+import {useI18n} from 'vue-i18n'
+import {useStore} from 'vuex'
 
 definePage({
   meta: {
-    title: "Perms",
+    title: 'Perms',
     requiresAuth: true
   }
-});
+})
 
-
-const { t } = useI18n()
+const {t} = useI18n()
 const store: Store = useStore()
 
 const sortBy = ref<SortBy[]>([{key: 'role', order: 'asc'}])
@@ -119,15 +101,15 @@ const dialog = ref(false)
 const selectedItem = ref<Partial<Permission> | null>(null)
 const search = ref('')
 
-const headers = ref<{title: string, key: keyof Permission | 'actions', info?: string | string[], maxWidth?: number}[]>([
-  { title: t('Role'), key: 'match'},
-  { title: t('Scopes'), key: 'scopes'},
-  { title: t('Actions'), key: 'actions', maxWidth: 20 }
+const headers = ref<{title: string; key: keyof Permission | 'actions'; info?: string | string[]; maxWidth?: number}[]>([
+  {title: t('Role'), key: 'match'},
+  {title: t('Scopes'), key: 'scopes'},
+  {title: t('Actions'), key: 'actions', maxWidth: 20}
 ])
 const items = computed(() => store.state.perms.items)
 const refresh = computed(() => store.state.refresh)
 
-watch(refresh, (val) => {
+watch(refresh, val => {
   if (!val) return
   getItems()
 })
@@ -144,19 +126,17 @@ function getScopes() {
 }
 
 function deleteItem(item: Permission) {
-  confirm(t('ConfirmDelete')) &&
-  store.dispatch('perms/deletePerm', item.id!)
+  confirm(t('ConfirmDelete')) && store.dispatch('perms/deletePerm', item.id!)
   getItems()
 }
 
 function editItem(item: Permission) {
-  selectedItem.value =  item
+  selectedItem.value = item
   dialog.value = true
 }
 
-
 function copyItem(item: Permission) {
-  selectedItem.value =  {...item, id: undefined}
+  selectedItem.value = {...item, id: undefined}
   dialog.value = true
 }
 
@@ -164,7 +144,6 @@ function close() {
   dialog.value = false
   selectedItem.value = null
 }
-
 
 getItems()
 getScopes()
